Move Neutron extractor outside nova
