+++ setup.cfg
@@ -48,7 +48,8 @@ console_scripts =
     caso-extract = caso.cmd.extract:main
 
 caso.extractors =
-    nova = caso.extract.openstack.nova:OpenStackExtractor
+    nova = caso.extract.openstack.nova:NovaExtractor
+    neutron = caso.extract.openstack.neutron:NeutronExtractor
 
 caso.messenger = 
     noop = caso.messenger.noop:NoopMessenger
